requirements_api.txt
# Core FastAPI service dependencies
fastapi==0.115.7
uvicorn==0.34.0
python-multipart==0.0.20

# Video/Audio processing
moviepy==2.2.1
torch==2.7.1
transformers==4.53.2
librosa==0.10.2
timm==1.0.11

# Text-to-Speech (edge-tts only)
edge-tts>=7.0.0

# Database and async support
sqlalchemy>=2.0.0
aiofiles>=24.0.0
httpx>=0.28.0
PyJWT>=2.10.0
jinja2>=3.1.0
python-dotenv>=1.1.0

# Web scraping (minimal set)
requests>=2.32.0
beautifulsoup4>=4.13.0
